Add retry button
We decided not to use polling and check internet access when app is being resumed. This button is added as a safety measure when device is connected to a network that has no internet access at the time of displaying flushbar and later regained the internet access on the same network.