Optimize swap polling, minimize SDK / mempool calls
The previous behavior sometimes lead to inconsistent
behavior, where the "Receive to BTC Address" UI would
show different swap addresses in short succession.

This commit fixes that, keeping the UI consistent by
showing the first unused swap address received. As
an additional benefit, it also reduces the amount of
calls to the SDK and to the mempool by half (from 2x
of each every 5 seconds, to 1x).